+++ R/mod_growth_death_rate.R
@@ -8,25 +8,25 @@
 #'
 #' @importFrom shiny NS tagList
 #' @importFrom shinycssloaders withSpinner
-mod_growth_death_rate_ui <- function(id, n_highligth = 5){
+mod_growth_death_rate_ui <- function(id){
   ns <- NS(id)
   tagList(
     fluidRow(
       column(6,
              uiOutput(ns("title_growth_factor")),
-             radioButtons(inputId = ns("growth_factor"), label = "",
+             selectInput(inputId = ns("growth_factor"), label = "",
                           choices = list("Over 3 days" = "growth_factor_3",
                                          "Over 5 days" = "growth_factor_5",
                                          "Over one week" = "growth_factor_7"),
-                          selected = "growth_factor_3", inline = TRUE),
+                          selected = "growth_factor_3"),
              withSpinner(uiOutput(ns("plot_growth_factor")))
       ),
       column(6,
-             uiOutput(ns("title_dath_toll")),
-             radioButtons(inputId = ns("radio_pop"), label = "",
+             uiOutput(ns("title_death_toll")),
+             selectInput(inputId = ns("radio_pop"), label = "",
                           choices = list("lethality rate" = "lethality_rate",
                                          "mortality rate 1M pop" = "mortality_rate_1M_pop"),
-                          selected = "lethality_rate", inline = TRUE),
+                          selected = "lethality_rate"),
              withSpinner(uiOutput(ns("plot_death_rate")))
       )
     )
@@ -81,27 +81,26 @@ mod_growth_death_rate_server <- function(input, output, session, df, n = 1000, w
   }
 
   # Dataset ----
-
   df_pop <- reactive(scale_mortality_rate(df))
 
-  df_base_plot1 <- reactive({pick_rate_hist( req(df_pop()), input$growth_factor)})
-  df_base_plot2 <- reactive({pick_rate_hist( req(df_pop()), input$radio_pop)})
+  df_base_plot1 <- reactive({pick_rate_hist( req(df_pop()), req(input$growth_factor))})
+  df_base_plot2 <- reactive({pick_rate_hist( req(df_pop()), req(input$radio_pop))})
 
   # Plots ----
 
   # titles
   if (istop) {
     output$title_growth_factor <- renderUI(div(h4(paste0("Current top ", n_highligth, " countries growth factor")), align = "center", style = "margin-top:20px; margin-bottom:20px;"))
-    output$title_dath_toll <- renderUI(div(h4(paste0("Current top ", n_highligth, " countries death toll")), align = "center", style = "margin-top:20px; margin-bottom:20px;"))
+    output$title_death_toll <- renderUI(div(h4(paste0("Current top ", n_highligth, " countries death toll")), align = "center", style = "margin-top:20px; margin-bottom:20px;"))
   } else {
     output$title_growth_factor <- renderUI(div(h4("Growth factor"), align = "center", style = "margin-top:20px; margin-bottom:20px;"))
-    output$title_dath_toll <- renderUI(div(h4("Death toll"), align = "center", style = "margin-top:20px; margin-bottom:20px;"))
+    output$title_death_toll <- renderUI(div(h4("Death toll"), align = "center", style = "margin-top:20px; margin-bottom:20px;"))
   }
 
   # captions
-  caption_growth_factor <- reactive({paste0("Computed as total confirmed cases today / total confirmed cases ", gsub("growth_factor_", "", input$growth_factor) ," days ago.")})
+  caption_growth_factor <- reactive({paste0("Computed as total confirmed cases today / total confirmed cases ", gsub("growth_factor_", "", req(input$growth_factor)) ," days ago.")})
   caption_death_rate_radio <- reactive({
-    if (input$radio_pop == "lethality_rate") {
+    if (req(input$radio_pop) == "lethality_rate") {
       p <- "/ total confirmed cases today."
     } else {
       p <- "per 1 M population "
@@ -109,22 +108,22 @@ mod_growth_death_rate_server <- function(input, output, session, df, n = 1000, w
     p
   })
   caption_death_rate <- reactive({paste0("Computed as total deaths today ",caption_death_rate_radio())})
-  caption_countries <- paste0("More than ", n, " confirmed cases and outbreaks longer than ", w, " days.")
+  #caption_countries <- paste0("More than ", n, " confirmed cases and outbreaks longer than ", w, " days.")
 
   # plots
   output$plot_growth_factor <- renderUI({
     tagList(
       plotlyOutput(ns("plot_growth_factor_hist"), height = 400),
-      div(p(caption_growth_factor()), align = "center"),
-      div(p(caption_countries), align = "center"),
+      div(p(caption_growth_factor()), align = "center")#,
+      #div(p(caption_countries), align = "center"),
     )
   })
 
   output$plot_death_rate <- renderUI({
     tagList(
       plotlyOutput(ns("plot_death_rate_hist"), height = 400),
-      div(p(caption_death_rate()), align = "center"),
-      div(p(caption_countries), align = "center"),
+      div(p(caption_death_rate()), align = "center")#,
+      #div(p(caption_countries), align = "center"),
     )
   })
 
@@ -139,9 +138,10 @@ mod_growth_death_rate_server <- function(input, output, session, df, n = 1000, w
     p
   })
 
-  is_percent <- reactive({ifelse(input$radio_pop == "lethality_rate", T, F)})
-
+  is_percent <- reactive({ifelse(req(input$radio_pop) == "lethality_rate", T, F)})
   output$plot_death_rate_hist <- renderPlotly({
+    message("is_percent = ", is_percent())
+
     g_palette_drf = g_palette[["death_rate"]]
 
     p <- plot_rate_hist(df_base_plot2(), percent = is_percent(), g_palette =  g_palette_drf)