+++ man-roxygen/ex-country.R
@@ -14,13 +14,12 @@ if (interactive()) {
 
   long_title <- "Lorem ipsum dolor sit amet, consectetur adipisicing elit."
   ui <- fluidPage(
-    Covid19Mirai:::mod_country_ui("country")
+    mod_country_ui("country")
   )
   server <- function(input, output) {
 
     orig_data <- get_datahub() %>%
       get_timeseries_by_contagion_day_data()
-
 
     pop_data = get_pop_datahub()
     orig_data_aggregate = build_data_aggr(orig_data, pop_data)
@@ -38,7 +37,7 @@ if (interactive()) {
     })
 
     callModule(mod_country_server, "country",
-               data_filtered = data_filtered, countries = countries)
+               data = data_filtered, countries = countries, n = n,  w = w)
   }
   runApp(shinyApp(ui = ui, server = server), launch.browser = TRUE)
 }